requirements.txt
transformers
pytorch-metric-learning[with-hooks-cpu]
